adding icon to test2
