+++ projects/test/src/main/java/com/example/test/HelloApplication.java
@@ -3,6 +3,7 @@ package com.example.test;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 
 import java.io.IOException;
