+++ projects/test/src/main/java/module-info.java
@@ -3,6 +3,7 @@ module com.example.test {
     requires javafx.fxml;
 
     requires org.kordamp.bootstrapfx.core;
+    requires javafx.swt;
 
     opens com.example.test to javafx.fxml;
     exports com.example.test;
