--- projects/test2/src/main/java/com/example/test2/HelloController.java
@@ -1,14 +1,0 @@
-package com.example.test2;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
