--- projects/test2/src/main/java/module-info.java
@@ -1,9 +1,0 @@
-module com.example.test2 {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-    requires org.kordamp.bootstrapfx.core;
-
-    opens com.example.test2 to javafx.fxml;
-    exports com.example.test2;
-}
