+++ projects/test2/src/main/module-info.java
@@ -1,0 +1,8 @@
+module test2 {
+    requires javafx.controls;
+    requires javafx.fxml;
+    requires org.kordamp.bootstrapfx.core;
+
+    opens test2 to javafx.fxml;
+    exports test2;
+}
